# public/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "StatusHub API",
    "version": "1.0.0",
    "description": "API REST para monitorización de servicios IT en tiempo real.",
    "contact": {
      "email": "admin@statushub.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Servidor local (Docker)"
    }
  ],
  "components": {
    "securitySchemes": {
      "sanctum": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token",
        "description": "Introduce el token obtenido en /api/auth/login"
      }
    }
  },
  "paths": {
...
    "/api/services/{service}": {
      "get": {
        "tags": ["Servicios"],
        "summary": "Detalle de un servicio",
        "description": "Devuelve los detalles de un servicio incluyendo sus incidentes activos.",
        "parameters": [
          {
            "name": "service",
            "in": "path",
            "required": true,
            "description": "ID del servicio",
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "responses": {
          "200": { "description": "Detalle del servicio obtenido correctamente" },
          "404": { "description": "Servicio no encontrado" }
        }
      }
    },
    "/api/services/{service}/status": {
      "put": {
        "tags": ["Servicios"],
        "summary": "Actualizar estado de un servicio",
        "description": "Actualiza manualmente el estado de un servicio. Solo administradores.",
        "security": [{ "sanctum": [] }],
        "parameters": [
          {
            "name": "service",
            "in": "path",
            "required": true,
            "description": "ID del servicio",
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["status"],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": ["operational", "degraded", "partial_outage", "major_outage"],
                    "example": "degraded"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Estado actualizado correctamente" },
          "401": { "description": "No autenticado" },
          "403": { "description": "No autorizado" }
        }
      }
    },
...
    "/api/subscriptions": {
      "post": {
        "tags": ["Suscripciones"],
        "summary": "Crear suscripción",
        "description": "Suscribe un email a las notificaciones de un servicio.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["service_id", "email"],
                "properties": {
                  "service_id": { "type": "integer", "example": 1 },
                  "email": { "type": "string", "format": "email", "example": "test@example.com" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Suscripción creada correctamente" },
          "409": { "description": "Ya existe una suscripción activa para este email y servicio" }
        }
      }
    },
    "/api/subscriptions/{token}": {
      "delete": {
        "tags": ["Suscripciones"],
        "summary": "Cancelar suscripción",
        "description": "Cancela una suscripción utilizando su token único.",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "description": "Token único de la suscripción",
            "schema": { "type": "string", "example": "YicWR5SiYfdf0FrkjuU1uFYDu2ZAmMbR" }
          }
        ],
        "responses": {
          "200": {
            "description": "Suscripción cancelada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Suscripción cancelada correctamente." }
                  }
                }
              }
            }
          },
          "404": { "description": "Suscripción no encontrada" }
        }
      }
    }
  }
}